+++ config/config.exs
@@ -1,5 +1,10 @@
-use Mix.Config
+import Config
 
 config :ex_correios,
-  base_url:
-    System.get_env("CORREIOS_URL") || "http://ws.correios.com.br/calculador/CalcPrecoPrazo.aspx"
+  calculator_url:
+    System.get_env("CORREIOS_CALCULATOR_URL") ||
+      "http://ws.correios.com.br/calculador/CalcPrecoPrazo.aspx"
+
+config :correios_cep, client: Correios.CEP.Client
+
+import_config "#{Mix.env()}.exs"