+++ config/test.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :correios_cep, client: ExCorreios.CEP.ClientStub