+++ lib/ex_correios/calculator/request/url.ex
@@ -1,19 +1,20 @@
-defmodule ExCorreios.Request.Url do
+defmodule ExCorreios.Calculator.Request.Url do
   @moduledoc """
   This module builds an url based on shipping informations.
   """
 
   @default_return "xml"
 
-  alias ExCorreios.Shipping.Shipping
+  alias ExCorreios.Calculator.Shipping
 
   @spec build(%Shipping{}, String.t()) :: String.t()
-  def build(shipping, base_url \\ nil), do: "#{base_url(base_url)}?#{format_params(shipping)}"
+  def build(shipping, calculator_url \\ nil),
+    do: "#{calculator_url(calculator_url)}?#{format_params(shipping)}"
 
   defp format_params(%{package: package} = shipping) do
     "nCdEmpresa=#{shipping.enterprise}&" <>
       "sDsSenha=#{shipping.password}&" <>
-      "nCdServico=#{format_services(shipping.services)}&" <>
+      "nCdServico=#{shipping.service.code}&" <>
       "sCepOrigem=#{shipping.origin}&" <>
       "sCepDestino=#{shipping.destination}&" <>
       "nVlPeso=#{package.weight}&" <>
@@ -30,14 +31,8 @@ defmodule ExCorreios.Request.Url do
       "StrRetorno=#{@default_return}"
   end
 
-  defp base_url(nil), do: Application.get_env(:ex_correios, :base_url)
-  defp base_url(url), do: url
-
-  defp format_services(services) do
-    services
-    |> Enum.map(fn {_k, v} -> v.code end)
-    |> Enum.join(",")
-  end
+  defp calculator_url(nil), do: Application.get_env(:ex_correios, :calculator_url)
+  defp calculator_url(url), do: url
 
   defp format_boolean(true), do: "S"
   defp format_boolean(false), do: "N"