+++ test/ex_correios/calculator/shipping_test.exs
@@ -1,9 +1,9 @@
-defmodule ExCorreios.Shipping.ShippingTest do
-  use ExUnit.Case
+defmodule ExCorreios.Calculator.ShippingTest do
+  use ExUnit.Case, async: true
 
   import ExCorreios.Factory
 
-  alias ExCorreios.Shipping.Shipping
+  alias ExCorreios.Calculator.Shipping
 
   describe "Shipping.new/3" do
     test "returns a shipping struct" do