+++ test/support/factory.ex
@@ -3,8 +3,8 @@ defmodule ExCorreios.Factory do
 
   use ExMachina
 
-  alias ExCorreios.Shipping.Packages.Package
-  alias ExCorreios.Shipping.{Service, Shipping}
+  alias ExCorreios.Calculator.Shipping
+  alias ExCorreios.Calculator.Shipping.{Package, Service}
 
   def package_factory do
     %Package{diameter: 40, format: 1, width: 11.0, height: 2.0, length: 16.0, weight: 0.3}
@@ -20,7 +20,7 @@ defmodule ExCorreios.Factory do
       package: package_factory(),
       password: "",
       receiving_alert: false,
-      services: Service.get_services([:pac])
+      service: Service.get_service(:pac)
     }
   end
 end