+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `ExCorreios.find_address/1` for search an address by a given postal code.
+
+### Fixed
+- Fix multiple services calculation.
+
+### Changed
+- Improve `.gitlab-ci.yml`.
 
 ## [1.0.0] - 2019-09-06
 ### Added